+++ block-accessibility-checks.php
@@ -2,31 +2,46 @@
 
 /**
  * Plugin Name:       Block Accessibility Checks
- * Description:       This plugin helps ensures your content meets WCAG (Web Content Accessibility Guidelines) requirements.
- * Requires at least: 6.3
- * Requires PHP:      7.0
+ * Plugin URI:        https://example.com/block-accessibility-checks
+ * Description:       This plugin helps ensure your content meets WCAG (Web Content Accessibility Guidelines) requirements.
  * Version:           0.1.0
+ * Requires at least: WordPress 6.3
+ * Requires PHP:      7.0
  * Author:            Troy Chaplin
+ * Author URI:        https://example.com
  * License:           GPL-2.0-or-later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain:       block-accessibility-checks
+ * Domain Path:       /languages
  *
  * @package           block-accessibility-checks
  */
 
-add_action('enqueue_block_editor_assets', 'enqueue_block_checks');
-
-function enqueue_block_checks()
-{
-    $script_path = 'build/index.js';
-
-    wp_register_script(
-        'block-check-script',
-        plugins_url($script_path, __FILE__),
-        array('wp-blocks', 'wp-i18n', 'wp-element', 'wp-editor'),
-        filemtime(plugin_dir_path(__FILE__) . $script_path),
-        true
-    );
-
-    wp_enqueue_script('block-check-script');
+// If this file is called directly, abort.
+if (!defined('ABSPATH')) {
+    die;
 }
+
+// Setup autoloading
+require_once __DIR__ . '/vendor/autoload.php';
+
+// Define constants
+if (!defined('BLOCK_ACCESSIBILITY_MODE')) {
+    define('BLOCK_ACCESSIBILITY_MODE', 'DENY'); // Default value, can be overridden with WARN in wp-config.php
+}
+
+// Include dependencies
+use BlockAccessibility\ScriptsAndStyles;
+use BlockAccessibility\Translations;
+
+// Define plugin file and Text Domain
+$pluginFile = __FILE__;
+$textDomain = 'block-accessibility-checks';
+
+// Translation setup
+$translations = new Translations($pluginFile, $textDomain);
+add_action('plugins_loaded', [$translations, 'loadTextDomain']);
+
+// Enqueue block editor assets
+$scriptsStyles = new ScriptsAndStyles($pluginFile, $translations);
+add_action('enqueue_block_editor_assets', [$scriptsStyles, 'enqueueAssets']);