+++ src/script.js
@@ -1,0 +1,7 @@
+import './styles/error.css';
+
+// Block Modifications
+import './scripts/blockMods/imageAttr';
+
+// Block Checks
+import './scripts/registerPlugin';