+++ src/scripts/helpers/blockChecksArray.js
@@ -1,3 +1,0 @@
-import { checkHeadingLevel } from '../blockChecks/checkHeading'
-
-export const blockChecksArray = [checkHeadingLevel]