+++ src/scripts/helpers/blockInvalidation.js
@@ -1,30 +1,42 @@
-import { useDispatch } from '@wordpress/data'
-import { getInvalidBlocks } from './getInvalidBlocks'
-import { useEffect } from '@wordpress/element'
+/* global blockAccessibilitySettings */
+import { useDispatch } from '@wordpress/data';
+import { GetInvalidBlocks } from './getInvalidBlocks';
+import { useEffect } from '@wordpress/element';
 
-export function blockInvalidation() {
-  const invalidBlocks = getInvalidBlocks()
+export function BlockInvalidation() {
+	const invalidBlocks = GetInvalidBlocks();
 
-  const {
-    lockPostSaving,
-    unlockPostSaving,
-    lockPostAutosaving,
-    unlockPostAutosaving,
-    disablePublishSidebar,
-    enablePublishSidebar,
-  } = useDispatch('core/editor')
+	const {
+		lockPostSaving,
+		unlockPostSaving,
+		lockPostAutosaving,
+		unlockPostAutosaving,
+		disablePublishSidebar,
+		enablePublishSidebar,
+	} = useDispatch('core/editor');
 
-  useEffect(() => {
-    if (invalidBlocks.length > 0) {
-      lockPostSaving()
-      lockPostAutosaving()
-      disablePublishSidebar()
-    } else {
-      unlockPostSaving()
-      unlockPostAutosaving()
-      enablePublishSidebar()
-    }
-  }, [invalidBlocks])
+	useEffect(() => {
+		if (
+			invalidBlocks.length > 0 &&
+			blockAccessibilitySettings.mode === 'DENY'
+		) {
+			lockPostSaving();
+			lockPostAutosaving();
+			disablePublishSidebar();
+		} else {
+			unlockPostSaving();
+			unlockPostAutosaving();
+			enablePublishSidebar();
+		}
+	}, [
+		invalidBlocks,
+		disablePublishSidebar,
+		enablePublishSidebar,
+		lockPostAutosaving,
+		lockPostSaving,
+		unlockPostAutosaving,
+		unlockPostSaving,
+	]);
 
-  return null
+	return null;
 }